requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
Automat==24.8.1
awscli==1.38.13
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
blis==1.2.0
boto==2.49.0
boto3==1.37.13
botocore==1.37.13
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
constantly==23.10.4
cryptography==43.0.3
cssselect==1.3.0
curl_cffi==0.9.0
cymem==2.0.11
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docutils==0.16
durationpy==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
fabric==3.2.2
fastapi==0.115.11
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.1
gotrue==2.11.4
groq==0.19.0
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
incremental==24.7.2
invoke==2.2.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
keybert==0.9.0
kubernetes==32.0.1
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.2.1
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.66.2
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parsel==1.10.0
pillow==11.1.0
postgrest==0.19.3
posthog==3.21.0
preshed==3.0.9
prettytable==3.15.1
propcache==0.3.0
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
PyDrive==1.3.1
PyDrive2==1.21.3
Pygments==2.19.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
queuelib==1.7.0
realtime==2.4.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.7.2
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
Scrapy==2.12.0
selenium==4.29.0
sentence-transformers==3.4.1
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.6.0
tldextract==5.1.3
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
w3lib==2.3.1
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
Wikipedia-API==0.8.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
